Add random string generator method and update stubs
